Fixed to compile.

+++ copasi/bindings/cpp_examples/example10/example10.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/bindings/cpp_examples/example10/example10.cpp,v $
-//   $Revision: 1.1.2.2 $
+//   $Revision: 1.1.2.3 $
 //   $Name:  $
-//   $Author: gauges $
-//   $Date: 2010/11/23 08:33:44 $
+//   $Author: shoops $
+//   $Date: 2011/01/12 21:43:44 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -259,7 +259,7 @@ int main(int argc, char** argv)
           // object is a vector with two unsigned int elements
           // First element is the index for the outer dimension and the second element is the index
           // for the inner dimension
-          std::vector<unsigned int> index(2);
+          std::vector<size_t> index(2);
 
           // since the rows and columns have the same annotation for the flux control coefficients, it doesn't matter
           // for which dimension we get the annotations
